add host
